+++ Particle_Transport/Material/material.cpp
@@ -10,20 +10,20 @@ short int Material::getAMass() {
 		return atomicMass;
 }
 
-double Material::getDensity() {
+float Material::getDensity() {
 		return density;
 }
 
-double Material::getADensity() {
+float Material::getADensity() {
 		return atomicDensity;
 }
 
-double Material::getXS(EventType et, const Particle& incP) {
+float Material::getXS(EventType et, const Particle& incP) {
 		return xstable.findRecord(et, incP).xs;
 }
 
-std::array<double, 3> Material::getEventXSs(const Particle& incP) {
-		std::array<double, 3> xss;
+std::array<float, 3> Material::getEventXSs(const Particle& incP) {
+		std::array<float, 3> xss;
 		std::array<XSRecord, 3> records = xstable.findEventRecords(incP);
 		xss[0] = records[0].xs;
 		xss[1] = records[1].xs;
