+++ Particle_Transport/Vector3/vector3.cpp
@@ -1,3 +1,5 @@
+// vector3.cpp
+
 #include "vector3.h"
 
 Vector3 Vector3::operator+(const Vector3& v) const {
@@ -8,11 +10,11 @@ Vector3 Vector3::operator-(const Vector3& v) const {
     return Vector3(x - v.x, y - v.y, z - v.z);
 }
 
-Vector3 Vector3::operator*(double a) const {
+Vector3 Vector3::operator*(float a) const {
     return Vector3(x * a, y * a, z * a);
 }
 
-double Vector3::dot(const Vector3& v) const {
+float Vector3::dot(const Vector3& v) const {
     return x * v.x + y * v.y + z * v.z;
 }
 
@@ -25,13 +27,13 @@ Vector3 Vector3::operator/(const Vector3& v) const {
 }
 
 void Vector3::normalize() {
-    double magnitude = std::sqrt(x*x + y*y + z*z);
+    float magnitude = std::sqrt(x*x + y*y + z*z);
     x /= magnitude;
     y /= magnitude;
     z /= magnitude;
 }
 
-Vector3 Vector3::randomOrthogonalV3(double u_x, double u_y, double u_z) const {
+Vector3 Vector3::randomOrthogonalV3(float u_x, float u_y, float u_z) const {
     // Create random vector with each component in [-1; 1] range (initial range [0; 1))
     Vector3 vec3(u_x * 2 - 1, u_y * 2 - 1, u_z * 2 - 1);
     // Make it orthogonal to current object
